Utility: * WIP commit for EDAN verifier * moved EDAN verifier into own class derived from VerifierBase * cleaned up how EDAN units are collected * opting for comparing abbreviations when using fetchFromNamedSearch and receiving multiple hits
