+++ server/tests/collections/EdanCollection.test.ts
@@ -113,7 +113,6 @@ describe('Collections: EdanCollection', () => {
     }
 });
 
-
 function executeTestQuery(ICol: COL.ICollection, query: string, expectNoResults: boolean,
     searchCollections: boolean = true, edanRecordType: string = ''): void {
     test('Collections: EdanCollection.queryCollection ' + query, async () => {
@@ -2244,737 +2243,24 @@ async function handle3DContentQuery(ICol: COL.ICollection, _WS: NodeJS.WritableS
 
 //----------------------------------------------------------------
 // #region EDAN Verifier
-// specific EDAN verifier
-import * as DBAPI from '../../db';
-import * as CACHE from '../../cache';
-import * as COMMON from '@dpo-packrat/common'; // shared by client & server
+import * as V from '../../utils/verifiers/EdanVerifier';
 
 function executeEdanVerifier(ICol: COL.ICollection) {
 
     // bad: don't keep
     test.only('Collections: EdanCollection.fieldCompare', async () => {
-        const result = await executeVerifySubjectGuts(ICol);
+        // const result = await executeVerifySubjectGuts(ICol);
+        const verifier: V.EdanVerifier = new V.EdanVerifier(ICol);
+        verifier.config.detailedLogs = false;
+        verifier.config.subjectLimit = 500;
+        verifier.config.systemObjectId = 463;
+        verifier.config.writeToFile = '../../EDAN-Verifier_Output.csv';
+        const result: boolean = await verifier.verify();
         expect(result).toBeTruthy();
     });
 
     //return true;
 }
-
-// [DPO3DPKRT-698]
-// TODO: ocassional failed fetch requests (ECONNRESET). make more robust
-// TODO: verify EDAN scene UUID are correct
-// TODO: verify EDAN 3d_package match the most recent published version
-// TODO: migrate into module to be called elsewhere
-// TODO: hook to endpoint and support running on a schedule via node_scheduler
-async function executeVerifySubjectGuts(ICol: COL.ICollection): Promise<boolean> {
-    const errorFn: string = 'EDAN Verifier'; //'EdanCollection.test.executeVerifySubjects';
-    const debugLogs: boolean = false;
-    const doFixErrors: boolean = false;
-    const debugSubjectLimit = 10000; // # of subjects to investigate for testing
-    const debugSysObjectID = -1;
-    jest.setTimeout(3000000); // temp: avoid timeout errors when doing a full run
-
-    // our structure and header for saved output to CSV
-    const output: string[] = [];
-    output.push('ID,MDM,URL,SUBJECT,STATUS,TEST,DESCRIPTION,PACKRAT,EDAN,NOTES');
-
-    // fetch all subjects from Packrat DB
-    const subjects: DBAPI.Subject[] | null = await DBAPI.Subject.fetchAll(); /* istanbul ignore if */
-    if (!subjects) {
-        LOG.error(`${errorFn} could not get subjects from DB`, LOG.LS.eTEST);
-        return false;
-    }
-    if(subjects.length<=0) {
-        LOG.error(`${errorFn} no subjects found in DB`, LOG.LS.eTEST);
-        return false;
-    }
-    if(debugLogs) console.log(`Subjects: ${subjects.length}`);
-
-    // loop through subjects, extract name, query from EDAN
-    for(let i=0; i<subjects.length; i++) {
-
-        // adjust our counter for # of subjects and bail when done
-        if(i>=debugSubjectLimit) break;
-
-        // create our stats object and helper variable
-        const subject = subjects[i];
-
-        // get our system object for the subject to help with logging and identification
-        const systemObject: DBAPI.SystemObject | null = await subject.fetchSystemObject();
-        if(!systemObject){
-            LOG.error(`could not get SystemObject from subject. (id: ${subject.idSubject} | subject: ${subject.Name})`, LOG.LS.eTEST);
-            continue;
-        }
-        if(debugLogs) console.log('SystemObject: '+JSON.stringify(systemObject));
-
-        // (debug) if not the desired subject id skip
-        if(debugSysObjectID>0){
-            if(systemObject.idSystemObject!==debugSysObjectID) {
-                if(debugLogs) console.log('Subject skipping. IDs don not match.');
-                continue;
-            }
-        }
-
-        LOG.info(`${errorFn} processing subject: ${subject.Name}`, LOG.LS.eTEST);
-        if(debugLogs) console.log('Subject:\t'+JSON.stringify(subject,null,0)); // temp
-
-        // get our subject's unit from Packrat DB
-        const packratUnit: DBAPI.Unit | null = await getSubjectUnit(subject);
-        if(!packratUnit) {
-            LOG.error(`${errorFn} could not find a unit for subject. skipping... (id: ${systemObject.idSystemObject} | subject: ${subject.Name})`, LOG.LS.eTEST);
-            continue;
-        }
-        if(debugLogs) console.log('Packrat Units: '+JSON.stringify(packratUnit,null,0));
-
-        // get our subject's identifiers, details, and SystemObject id
-        const packratIdentifiers: IdentifierList | null = await getSubjectIdentifiers(subject,systemObject);
-        if(!packratIdentifiers) {
-            LOG.error(`${errorFn} could not get identifiers for subject. skipping... (id: ${systemObject.idSystemObject} | subject: ${subject.Name})`, LOG.LS.eTEST);
-            continue;
-        }
-        if(debugLogs) console.log('Packrat Identifiers: '+JSON.stringify(packratIdentifiers,null,0));
-
-        // if we have an EDAN (should 100% be true) then determine if from DPO or not
-        // defaults to all subjects coming from EDAN
-        let isDPOSubject: boolean = false;
-        if(packratIdentifiers.edan && packratIdentifiers.edan.identifier) {
-            isDPOSubject = isIdentifierFromDPO(packratIdentifiers.edan.identifier);
-        } else {
-            LOG.error(`${errorFn} could not get EDAN ID for subject. source of subject is ambiguous. (id: ${systemObject.idSystemObject} | subject: ${subject.Name})`, LOG.LS.eTEST);
-        }
-
-        // get our packrat name
-        const packratName: string = subject.Name;
-
-        // if we have an EDAN id we use it
-        let query: string | undefined = '';
-        if(packratIdentifiers.edan?.identifier)
-            query = packratIdentifiers.edan.identifier.IdentifierValue;
-        else if(packratIdentifiers.ark?.identifier)
-            query = packratIdentifiers.ark.identifier.IdentifierValue;
-        else if(packratIdentifiers.preferred?.identifier)
-            query = packratIdentifiers.preferred.identifier.IdentifierValue;
-        else {
-            LOG.error(`${errorFn} no good options for querying EDAN for subject. skipping... (id: ${systemObject.idSystemObject} | subject: ${subject.Name})`, LOG.LS.eTEST);
-            continue;
-        }
-        if(debugLogs) console.log('Query: '+query);
-
-        // query EDAN with our identifier and skip if nothing returned
-        const options: COL.CollectionQueryOptions | null = {
-            recordType: 'edanmdm',
-            // gatherRaw: true,
-            gatherIDMap: true,
-        };
-        const results: COL.CollectionQueryResults | null = await ICol.queryCollection(query, 10, 0, options);
-        if(!results || results.records.length<=0) {
-            LOG.error(`${errorFn} did not receive records for subject and identifier. skipping... (id: ${systemObject.idSystemObject} | subject: ${subject.Name} | query: ${query})`, LOG.LS.eTEST);
-            continue;
-        }
-        if(debugLogs) console.log('EDAN Results: '+JSON.stringify(results,null,0));
-
-        // structure to hold our units/identifiers
-        let edanUnit: DBAPI.Unit | null = null;
-        let edanIdentifiers: IdentifierList | null = null;
-        let edanName: string = '';
-
-        // if we received multipled records from EDAN, complain but keep going (compares will place subject in output)
-        if(results.records.length==1) {
-            const record = results.records[0];
-
-            // get our EDAN units from this record
-            const unit: DBAPI.Unit | null = await getEdanRecordUnit(record,packratUnit);
-            if(!unit)
-                LOG.error(`${errorFn} no known units found for subject (id: ${systemObject.idSystemObject} | subject: ${subject.Name}) and EDAN record name (${record.name})`, LOG.LS.eTEST);
-            else
-                edanUnit = unit;
-            if(debugLogs) console.log('EDAN Units: '+JSON.stringify(unit,null,0));
-
-            // get our EDAN identifiers from this record
-            const identifiers: IdentifierList | null = await getEdanRecordIdentifiers(record);
-            if(!identifiers)
-                LOG.error(`${errorFn} no EDAN identifiers found for subject (id: ${systemObject.idSystemObject} | subject: ${subject.Name}) and EDAN record name (${record.name})`, LOG.LS.eTEST);
-            else
-                edanIdentifiers = identifiers;
-            if(debugLogs) console.log('EDAN Ids: '+JSON.stringify(identifiers,null,0));
-
-            // get our name
-            edanName = record.name;
-
-        } else {
-            LOG.error(`${errorFn} received multiple records from EDAN when expecting 1. skipping... (id: ${systemObject.idSystemObject} | subject: ${subject.Name} | query: ${query})`, LOG.LS.eTEST);
-        }
-
-        // a structure to hold our output
-        const outputPrefix = `${systemObject.idSystemObject},${packratIdentifiers.edan?.identifier?.IdentifierValue},${getSystemObjectDetailsURL(systemObject)},${JSON.stringify(subject.Name)},`;
-
-        // Compare: Name
-        if(packratName!=edanName) {
-            LOG.error(`${errorFn} Subject name in Packrat and EDAN are not the same (id:${systemObject.idSystemObject} | Packrat:${packratName} | EDAN:${edanName})`, LOG.LS.eTEST);
-            let str = outputPrefix;
-            str += 'error,';
-            str += 'name,';
-            str += 'Subject name not the same,';
-            str += '"' + packratName + '",';
-            str += '"' + edanName + '",';
-            str += ((isDPOSubject)?'DPO created subject. needs manual fix':'EDAN subject. needs manual fix.')+',';
-            output.push(str);
-        } else {
-            LOG.info(`${errorFn} name compare succeeded!`, LOG.LS.eTEST);
-            let str = outputPrefix;
-            str += 'success,';
-            str += 'name,';
-            str += ',';
-            str += '"' + packratName + '",';
-            str += '"' + edanName + '",';
-            str += ',';
-            output.push(str);
-        }
-
-        // Compare: Units
-        if(packratUnit && edanUnit) {
-            // if both present then we check the value match
-            let foundUnit: boolean = false;
-            if(edanUnit.idUnit===packratUnit.idUnit) {
-                foundUnit = true;
-            }
-
-            // if we couldn't find the unit, add error otherwise success
-            if(!foundUnit) {
-                LOG.error(`${errorFn} Packrat unit does not match EDAN units (id:${systemObject.idSystemObject} | subject:${subject.Name})`, LOG.LS.eTEST);
-                let str = outputPrefix;
-                str += 'error,';
-                str += 'units,';
-                str += 'Packrat unit does not match EDAN units,';
-                str += '"' + packratUnit.Abbreviation + ' - ' + packratUnit.Name + '",';
-                str += '"' + edanUnit.Abbreviation + ' - ' + edanUnit.Name +'",';
-
-                // apply edan unit packrat subject
-                if(isDPOSubject) {
-                    str += 'DPO subject. needs manual fix to update EDAN unit info';
-                } else {
-                    if(doFixErrors) {
-                        const replaceUnitResult: boolean = await replacePackratUnit(packratUnit,edanUnit);
-                        if(!replaceUnitResult) {
-                            LOG.error(`${errorFn} failed to update Packrat unit to match EDAN (id:${systemObject.idSystemObject} | subject:${subject.Name})`, LOG.LS.eTEST);
-                            str += 'EDAN subject. failed to automatic update. check logs or do manually';
-                        } else {
-                            LOG.info(`${errorFn} successfully updated Packrat unit to match EDAN (id:${systemObject.idSystemObject} | subject:${subject.Name})`, LOG.LS.eTEST);
-                            str += 'EDAN subject. updated Packrat unit to match EDAN';
-                        }
-                    } else {
-                        str += 'EDAN subject. automatic updating disabled. rerun or manual fix required';
-                    }
-                }
-
-                // store for output
-                output.push(str);
-            } else {
-                LOG.info(`${errorFn} Unit compare succeeded! (id:${systemObject.idSystemObject} | subject:${subject.Name})`, LOG.LS.eTEST);
-                let str = outputPrefix;
-                str += 'success,';
-                str += 'units,';
-                str += ',';
-                str += '"' + packratUnit.Abbreviation + ' - ' + packratUnit.Name + '",';
-                str += '"' + edanUnit.Abbreviation + ' - ' + edanUnit.Name + '",';
-                output.push(str);
-            }
-        } else if(packratUnit && !edanUnit) {
-            LOG.error(`${errorFn} Packrat unit not found in EDAN record (id:${systemObject.idSystemObject} | subject:${subject.Name} | packrat:${JSON.stringify(packratUnit)})`, LOG.LS.eTEST);
-            let str = outputPrefix;
-            str += 'error,';
-            str += 'units,';
-            str += 'Packrat unit not found in EDAN,';
-            str += '"' + packratUnit.Abbreviation + ' - ' + packratUnit.Name + '",';
-            str += 'null,';
-            str += ((isDPOSubject)?'DPO created subject. needs manual fix to apply to EDAN.':'EDAN subject. needs manual fix because EDAN should have unit assigned. (todo)')+',';
-            output.push(str);
-        } else if(!packratUnit && edanUnit) {
-            LOG.error(`${errorFn} EDAN unit not found in Packrat (id:${systemObject.idSystemObject} | subject:${subject.Name} | EDAN:${edanUnit.Name})`, LOG.LS.eTEST);
-            let str = outputPrefix;
-            str += 'error,';
-            str += 'units,';
-            str += 'EDAN unit not found in Packrat,';
-            str += 'null,';
-            str += '"' + edanUnit.Abbreviation + ' - ' + edanUnit.Name +'",';
-
-            // apply edan unit to packrat unit (creating new unit in process)
-            if(isDPOSubject) {
-                str += 'DPO subject. needs manual fix as Packrat should have unit';
-            } else {
-                if(doFixErrors) {
-                    const replaceUnitResult: boolean = await replacePackratUnit(packratUnit,edanUnit);
-                    if(!replaceUnitResult) {
-                        LOG.error(`${errorFn} failed to update Packrat unit to match EDAN (id:${systemObject.idSystemObject} | subject:${subject.Name})`, LOG.LS.eTEST);
-                        str += 'EDAN subject. failed to automatic update. check logs or do manually';
-                    } else {
-                        LOG.info(`${errorFn} successfully updated Packrat unit to match EDAN (id:${systemObject.idSystemObject} | subject:${subject.Name})`, LOG.LS.eTEST);
-                        str += 'EDAN subject. updated Packrat unit to match EDAN';
-                    }
-                } else {
-                    str += 'EDAN subject. automatic updating disabled. rerun or manual fix required';
-                }
-            }
-
-            // store for output
-            output.push(str);
-
-        } else {
-            LOG.error(`${errorFn} Packrat & EDAN units not found (id:${systemObject.idSystemObject} | subject:${subject.Name})`, LOG.LS.eTEST);
-            let str = outputPrefix;
-            str += 'error,';
-            str += 'units,';
-            str += 'Packrat and EDAN units not,';
-            str += 'null,';
-            str += 'null,';
-            str += 'needs manual fix as neither source has a unit,';
-            output.push(str);
-        }
-
-        // Compare: Identifiers
-        if(packratIdentifiers && edanIdentifiers) {
-            let idMismatch: boolean = false;
-
-            // figure out how many identifiers edan has
-            const edanIdCount = edanIdentifiers.details.length;
-
-            // build our list of identifiers
-            // TODO: clean strings for CSV output
-            const strPackratIds: string[] = (packratIdentifiers.details)?(packratIdentifiers.details?.map(id=>{ return id.identifier?.IdentifierValue+' ('+id.identifierType?.Term+')'; })):([]);
-            const strEdanIds: string[] = [];
-            for(const id of edanIdentifiers.details) {
-                strEdanIds.push(id.identifier?.IdentifierValue+' ('+id.identifierType?.Term+')');
-            }
-
-            // see if we have the same count for a quick catch of difference, otherwise do deeper compare
-            if(packratIdentifiers.details?.length != edanIdCount) {
-                idMismatch = true;
-            } else {
-                // cycle through all edan identifiers and look for match in packrat
-                let didFindId = false;
-                for(const id of edanIdentifiers.details) {
-                    for(const packratId of packratIdentifiers.details) {
-                        if(packratId.identifierTypeEnum===id.identifierTypeEnum &&
-                            packratId.identifier?.IdentifierValue===id.identifier?.IdentifierValue) {
-                            didFindId = true;
-                            break;
-                        }
-                    }
-                }
-                if(!didFindId) idMismatch = true;
-            }
-
-            // if we have a mismatch then output
-            if(idMismatch) {
-                // log error and details
-                LOG.error(`${errorFn} Packrat has different identifiers than EDAN (id:${systemObject.idSystemObject} | subject:${subject.Name})`, LOG.LS.eTEST);
-                LOG.error(`${errorFn} \tPackrat: `+strPackratIds.sort().join(','), LOG.LS.eTEST);
-                LOG.error(`${errorFn} \t   EDAN: `+strEdanIds, LOG.LS.eTEST);
-
-                // if we're an EDAN subject then we need to fix the situation
-                if(!isDPOSubject) {
-                    // TODO: wipe out Packrat identifiers and replace with EDAN modifiers
-                    const replaceIdsResult: boolean = await replacePackratIdentifiers(packratIdentifiers,edanIdentifiers,systemObject);
-                    if(!replaceIdsResult) console.error('failed');
-                }
-
-                // build our output string
-                let str = outputPrefix;
-                str += 'error,';
-                str += 'identifiers,';
-                str += 'Packrat identifiers do not match EDAN,';
-                str += '"'+strPackratIds.sort().join('\n')+'",';
-                str += '"'+strEdanIds.sort().join('\n')+'",';
-
-                // apply edan unit to packrat unit (creating new unit in process)
-                if(isDPOSubject) {
-                    str += 'DPO subject. needs manual fix';
-                } else {
-                    if(doFixErrors) {
-                        const replaceUnitResult: boolean = await replacePackratIdentifiers(packratIdentifiers,edanIdentifiers,systemObject);
-                        if(!replaceUnitResult) {
-                            LOG.error(`${errorFn} failed to update Packrat identifiers to match EDAN (id:${systemObject.idSystemObject} | subject:${subject.Name})`, LOG.LS.eTEST);
-                            str += 'EDAN subject. failed to automatic update. check logs or do manually';
-                        } else {
-                            LOG.info(`${errorFn} successfully updated Packrat identifiers to match EDAN (id:${systemObject.idSystemObject} | subject:${subject.Name})`, LOG.LS.eTEST);
-                            str += 'EDAN subject. updated Packrat identifiers to match EDAN';
-                        }
-                    } else {
-                        str += 'EDAN subject. automatic updating disabled. rerun or manual fix required';
-                    }
-                }
-
-                // store for output
-                output.push(str);
-
-            } else {
-                LOG.info(`${errorFn} Identifier compare succeeded! (id:${systemObject.idSystemObject} | subject:${subject.Name})`, LOG.LS.eTEST);
-                let str = outputPrefix;
-                str += 'success,';
-                str += 'identifiers,';
-                str += ',';
-                str += '"'+strPackratIds.sort().join('\n')+'",';
-                str += '"'+strEdanIds.sort().join('\n')+'",';
-                str += ',';
-                output.push(str);
-            }
-        } else if(packratIdentifiers && !edanIdentifiers) {
-            // TODO
-        } else if(!packratIdentifiers && edanIdentifiers) {
-            // TODO
-        }
-    }
-
-    // HACK: dumping to local file until moved into Workflow reports.
-    require('fs').writeFile('../../EDAN-Verifier_Output.csv',output.join('\n'), err=>{ if(err) { console.error(err); }});
-
-    return true;
-}
-
-type IdentifierDetails = {
-    identifier: DBAPI.Identifier | null,
-    identifierType: DBAPI.Vocabulary | null,
-    identifierTypeEnum: COMMON.eVocabularyID | null,
-};
-
-type IdentifierList = {
-    preferred: IdentifierDetails | null,
-    edan: IdentifierDetails | null,
-    ark: IdentifierDetails | null,
-    details: IdentifierDetails[],  // complete list of all identifiers
-};
-
-async function getIdentifierType(identifier: DBAPI.Identifier): Promise<IdentifierDetails | null> {
-
-    // grab the identifier type object from the Packrat DB
-    const identifierType: DBAPI.Vocabulary | null = await DBAPI.Vocabulary.fetch(identifier.idVIdentifierType);
-    if(!identifierType){
-        LOG.error(`could not find identifier type in DB (identifier: ${identifier.idVIdentifierType} )`, LOG.LS.eTEST);
-        return null;
-    }
-    // if(debugLogs) console.log('Type:\t'+JSON.stringify(identifierType)); //temp
-
-    // pull from enumeration from the CACHE (vocabulary id -> enum)
-    const identifierTypeEnum: COMMON.eVocabularyID | undefined = await CACHE.VocabularyCache.vocabularyIdToEnum(identifier.idVIdentifierType);
-    if(identifierTypeEnum===undefined){
-        LOG.error(`could not find enumerator for identifier type (${identifier.idVIdentifierType}) in Cache`, LOG.LS.eTEST);
-        return null;
-    }
-    // if(debugLogs) console.log('Enum:\t'+JSON.stringify(identifierTypeEnum,null,0));
-
-    return { identifier, identifierType, identifierTypeEnum };
-}
-
-async function getSubjectIdentifiers(subject: DBAPI.Subject, systemObject: DBAPI.SystemObject): Promise<IdentifierList | null> {
-
-    // structure to hold our results
-    const result: IdentifierList = { preferred: null, edan: null, ark: null, details: [] };
-
-    // grab the preferred identifier, if nothing then leave null so calling function can decide action
-    if(subject.idIdentifierPreferred) {
-        const preferredIdentifier: DBAPI.Identifier | null = await DBAPI.Identifier.fetch(subject.idIdentifierPreferred);
-        if(!preferredIdentifier){
-            LOG.error(`subject's preferredId not found in the DB (id: ${subject.idSubject} | subject: ${subject.Name})`, LOG.LS.eTEST);
-            // subjectStats[i].isValid = false;
-        } else {
-            // grab our identifier details (type) and store it
-            const preferredIdentifierDetails: IdentifierDetails | null = await getIdentifierType(preferredIdentifier);
-            if(preferredIdentifierDetails) {
-                result.preferred = preferredIdentifierDetails;
-                result.details?.push(preferredIdentifierDetails);
-            }
-
-            // check for edan/ark and store as appropriate because the preferred id may be ARK too
-            switch(preferredIdentifierDetails?.identifierTypeEnum){
-                case COMMON.eVocabularyID.eIdentifierIdentifierTypeEdanRecordID: {
-                    if(!result.edan) result.edan = preferredIdentifierDetails;
-                } break;
-
-                case COMMON.eVocabularyID.eIdentifierIdentifierTypeARK: {
-                    if(!result.ark) result.ark = preferredIdentifierDetails;
-                } break;
-            }
-        }
-    }
-
-    // grab our list of identifiers from the SystemObject id
-    const identifiers: DBAPI.Identifier[] | null = await DBAPI.Identifier.fetchFromSystemObject(systemObject.idSystemObject);
-    if(!identifiers){
-        LOG.error(`could not get identifiers from subject (id: ${subject.idSubject} | subject: ${subject.Name})`, LOG.LS.eTEST);
-        return null;
-    }
-    if(identifiers.length<=0){
-        LOG.info(`(WARNING) no identifiers assigned to subject (id: ${subject.idSubject} | subject: ${subject.Name})`, LOG.LS.eTEST);
-        return result;
-    }
-
-    // cycle through all identifiers, find EDAN/ARK, and push to list
-    for(const identifier of identifiers) {
-
-        // get our details for this identifier, skip if error, store if valid
-        const details = await getIdentifierType(identifier);
-        if(!details) {
-            LOG.error(`could not get identifier details from subject (identifier: ${identifier.IdentifierValue} | id: ${subject.idSubject} | subject: ${subject.Name})`, LOG.LS.eTEST);
-            continue;
-        }
-
-        // make sure it doesn't already exist before pushing it
-        let idExists: boolean = false;
-        for(const id of result.details){
-            if(details.identifierTypeEnum===id.identifierTypeEnum &&
-                details.identifier?.IdentifierValue==id.identifier?.IdentifierValue) {
-                idExists = true;
-            }
-        }
-        if(idExists) continue;
-
-        // if not found push it and categorize it
-        result.details?.push(details);
-
-        // check the enumeration type to see if it's an edan or ark type
-        switch(details.identifierTypeEnum){
-            case COMMON.eVocabularyID.eIdentifierIdentifierTypeEdanRecordID: {
-                if(!result.edan) result.edan = details;
-            } break;
-
-            case COMMON.eVocabularyID.eIdentifierIdentifierTypeARK: {
-                if(!result.ark) result.ark = details;
-            } break;
-        }
-    }
-
-    return result;
-}
-
-async function getSubjectUnit(subject: DBAPI.Subject): Promise<DBAPI.Unit | null> {
-
-    const packratUnit = await DBAPI.Unit.fetch(subject.idUnit);
-    if(!packratUnit) {
-        LOG.error(`Packrat DB did not return a unit for subject. (id: ${subject.idSubject} | subject: ${subject.Name})`, LOG.LS.eTEST);
-        return null;
-    }
-
-    // Todo: any additional verification or handling?
-
-    return packratUnit;
-}
-
-async function getEdanRecordUnit(record: COL.CollectionQueryResultRecord, packratUnit: DBAPI.Unit): Promise<DBAPI.Unit | null> {
-
-    // TODO: additional verification or handling? (e.g. create new Unit for unknown types?)
-    // TODO: relocate logic to central/shared location to benefit ingestion
-
-    // see if UnittEdan has a direct match for this unit, if not do a deeper (less precise approach)
-    const edanUnits: DBAPI.UnitEdan[] | null = await DBAPI.UnitEdan.fetchFromName(record.unit);
-    if(edanUnits && edanUnits.length==1 && edanUnits[0].idUnit) {
-        const result = DBAPI.Unit.fetch(edanUnits[0].idUnit);
-        if(result) return result;
-    }
-
-    LOG.info(`(WARNING) did not find EDAN unit in the UnitEdan DB. Doing less precise check... (${record.unit}) `, LOG.LS.eTEST);
-
-    // if not found then do a deeper named search on Packrat units
-    const packratUnits: DBAPI.Unit[] | null = await DBAPI.Unit.fetchFromNameSearch(record.unit);
-    if(!packratUnits){
-        LOG.info(`(WARNING) did not find known Packrat unit for EDAN unit. (${record.unit})`, LOG.LS.eTEST);
-        return null;
-    }
-
-    // make sure we didn't get multiples
-    if(packratUnits.length!=1) {
-        LOG.error(`received multiple units from EDAN. Should only be 1 for subjects. Checking if packrat unit is in list. (${record.unit}:${packratUnits.map(x=>{ return x.Abbreviation; }).join('|')})`, LOG.LS.eTEST);
-
-        // now we check if our unit from packrat is in the returned results
-        // TEST: if so, we return that instead...?
-        for(const unit of packratUnits)
-            if(unit.idUnit===packratUnit.idUnit) return unit;
-
-        return null;
-    }
-
-    // if wqe got multiples, then we need to throw error. It should only be one.
-    return packratUnits[0];
-}
-
-async function getEdanRecordIdentifiers(record: COL.CollectionQueryResultRecord): Promise<IdentifierList | null> {
-
-    // structure to hold our results
-    const result: IdentifierList = { preferred: null, edan: null, ark: null, details: [] };
-
-    // see if we have an EDAN id stored
-    if(record.identifierCollection) {
-        // get the identifier if it exists
-        // HACK: prefixing identifier with 'edanmdm' to match Packrat's records
-        const edanIdentifiers: DBAPI.Identifier[] | null = await DBAPI.Identifier.fetchFromIdentifierValue('edanmdm:'+record.identifierCollection);
-        if(edanIdentifiers) {
-
-            // cycle through and get our type and details
-            for(const identifier of edanIdentifiers) {
-                const details: IdentifierDetails | null = await getIdentifierType(identifier);
-                if(!details) {
-                    LOG.error(`could not get details for EDAN identifier (type: ${identifier.idVIdentifierType} | value:${identifier.IdentifierValue})`, LOG.LS.eTEST);
-                    continue;
-                }
-
-                // if we have an identifier that is of the same type then store
-                if(details.identifierTypeEnum===COMMON.eVocabularyID.eIdentifierIdentifierTypeEdanRecordID) {
-                    result.edan = details;
-                    result.details?.push(details);
-                    break;
-                }
-            }
-        } else {
-            // TODO: create new identifier, type, and of EDAN Record ID type
-        }
-    }
-
-    // see if we have an ARK id stored
-    if(record.identifierPublic) {
-        // get the identifier if it exists
-        const arkIdentifiers: DBAPI.Identifier[] | null = await DBAPI.Identifier.fetchFromIdentifierValue(record.identifierPublic);
-        if(arkIdentifiers) {
-            // cycle through and get our type and details
-            for(const identifier of arkIdentifiers) {
-                const details: IdentifierDetails | null = await getIdentifierType(identifier);
-                if(!details) {
-                    LOG.error(`could not get details for ARK identifier (type: ${identifier.idVIdentifierType} | value:${identifier.IdentifierValue})`, LOG.LS.eTEST);
-                    continue;
-                }
-
-                // if we have an identifier that is of the same type then store
-                // todo: verify it's an actual ARK id
-                if(details.identifierTypeEnum===COMMON.eVocabularyID.eIdentifierIdentifierTypeARK) {
-                    result.ark = details;
-                    result.details?.push(details);
-                    break;
-                }
-            }
-        } else {
-            // todo: create new identifier, type, and of ARK type
-        }
-    }
-
-    // handle identifiers by checking if any returned by EDAN
-    if(record.identifierMap) {
-
-        // console.log('EDAN: '+H.Helpers.JSONStringify(record.identifierMap));
-        for (const [ label, content ] of record.identifierMap) {
-
-            // get our type for this identifier
-            const identifierType: DBAPI.Vocabulary | undefined = await CACHE.VocabularyCache.mapIdentifierType(label);
-            if (!identifierType) {
-                LOG.error(`\tencountered unknown identifier type ${label} for EDAN record ${record.name}`, LOG.LS.eTEST);
-                continue;
-            }
-
-            // pull from enumeration from the CACHE (vocabulary id -> enum)
-            const identifierTypeEnum: COMMON.eVocabularyID | undefined = await CACHE.VocabularyCache.vocabularyIdToEnum(identifierType.idVocabulary);
-            if(identifierTypeEnum===undefined){
-                LOG.error(`\tcould not find enumerator for identifier type (${identifierType.Term}) in Cache`, LOG.LS.eTEST);
-                continue;
-            }
-
-            // if identifier exists in our database (value & type) then store it
-            const identifiers: DBAPI.Identifier[] | null = await DBAPI.Identifier.fetchFromIdentifierValue(content);
-            if(identifiers) {
-                for(const identifier of identifiers) {
-                    const details: IdentifierDetails | null = await getIdentifierType(identifier);
-                    if(!details) {
-                        LOG.error(`could not get details for EDAN identifier (type: ${identifier.idVIdentifierType} | value:${identifier.IdentifierValue})`, LOG.LS.eTEST);
-                        continue;
-                    }
-
-                    // if we have an identifier that is of the same type then store
-                    if(details.identifierTypeEnum===identifierTypeEnum) {
-                        result.details?.push(details);
-                        break;
-                    }
-                }
-
-            } else {
-                // didn't find the identifier in our database so create one
-                // TODO: make DBAPI.Identifier object
-
-                const details: IdentifierDetails = { identifier: null, identifierType, identifierTypeEnum };
-                result.details?.push(details);
-            }
-
-            // console.log('EDAN: '+label+'|'+content+'|'+JSON.stringify(vIdentifierType));
-        }
-    }
-
-    return result;
-}
-
-async function replacePackratUnit(packratUnit: DBAPI.Unit | null, edanUnit: DBAPI.Unit): Promise<boolean> {
-
-    // if packrat unit is null, we need to create one and push to the system
-    if(!packratUnit) { return false; }
-    //     console.warn(edanUnit);
-
-    //     const unitArgs = {
-    //         idUnit: 0, // use Edan's?
-    //         Name: edanUnit.Name,
-    //         Abbreviation: edanUnit.Abbreviation,
-    //         ARKPrefix: edanUnit.ARKPrefix
-    //     };
-    //     packratUnit = new DBAPI.Unit(unitArgs);
-    //     const wasCreated: boolean = await packratUnit.create();
-    //     if(!wasCreated) {
-    //         LOG.error(`failed to create new unit in Packrat (unit: ${edanUnit.Name})`, LOG.LS.eTEST);
-    //         return false;
-    //     }
-
-    //     LOG.info(`Created new unit in Packrat (id: ${packratUnit.idUnit}} | unit: ${packratUnit.Name})`, LOG.LS.eTEST);
-    //     return true;
-    // }
-
-    // copy properties over. skip the id or things will break
-    packratUnit.Name = edanUnit.Name;
-    packratUnit.Abbreviation = edanUnit.Abbreviation;
-    packratUnit.ARKPrefix = edanUnit.ARKPrefix;
-
-    const result: boolean =  await packratUnit.update();
-    if(!result) {
-        LOG.error(`could not update Packrat unit (id: ${packratUnit.idUnit}} | unit: ${packratUnit.Name})`, LOG.LS.eTEST);
-        return false;
-    }
-    return true;
-}
-
-async function replacePackratIdentifiers(_packratIdentifiers: IdentifierList, _edanIdentifiers: IdentifierList, _systemObject: DBAPI.SystemObject): Promise<boolean> {
-
-    // [?] do we remove previous identifiers?
-    // [?] do we repurpose them by reassign new values keeping ids (still need to add/remove if count mismatch)?
-
-    // get subject system object for identifiers
-
-    // wipe all identifiers from subject
-    // for(const id of packratIdentifiers.details) {
-    //     await id.identifier?.delete();
-    // }
-
-    // cycle through edan identifiers creating new entries in the DB for each attached to the same SystemObject
-
-    return true;
-}
-
-function isIdentifierFromDPO(identifier: DBAPI.Identifier): boolean {
-    // simply check if the EDAN id starts with the DPO prefix.
-    // TODO: make more robust with additional checks(?)
-    return (identifier.IdentifierValue.startsWith('dpo_3d') || identifier.IdentifierValue.startsWith('edanmdm:dpo_3d'));
-}
-
-/* eslint-disable @typescript-eslint/no-unused-vars */
-function getSystemObjectDetailsURL(systemObject: DBAPI.SystemObject) {
-    return '=HYPERLINK("https://packrat-test.si.edu:8443/repository/details/'+systemObject.idSystemObject+'")';
-}
-
-// function getSubjectDetailsURL(subject: DBAPI.Subject) {
-//     return 'https://packrat-test.si.edu:8443/repository/details/'+subject.idSubject;
-// }
-
-/* eslint-enable @typescript-eslint/no-unused-vars */
 
 // #endregion
